api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg2-binary==2.9.9
prometheus-client==0.20.0
mlflow==2.14.1
